Handle exception and error for the application accordingly
