+++ src/main/java/Duke/Deadline.java
@@ -1,4 +1,4 @@
-package Tasks;
+package Duke;
 
 public class Deadline extends Task {
