+++ src/main/java/Duke/DukeException.java
@@ -1,0 +1,14 @@
+package Duke;
+
+public class DukeException extends Exception {
+    
+    /**
+	 *
+	 */
+	private static final long serialVersionUID = -2154386125699489110L;
+
+	public DukeException(String message) {
+        super(message);
+    }
+
+}
